Extract repeated code in report generation
Many A&D reports have the same structure - organized by income interval,
then by Borrower Characteristics (race, ethnicity, and minority status),
then by disposition (application received, originated, denied,
withdrawn, etc.) this method organizes data into that structure.
