+++ publication/src/main/scala/hmda/publication/reports/disclosure/D51.scala
@@ -4,7 +4,6 @@ import java.util.Calendar
 
 import akka.NotUsed
 import akka.stream.scaladsl.Source
-import hmda.model.publication.reports.ApplicantIncomeEnum._
 import hmda.publication.reports._
 import hmda.model.publication.reports._
 import hmda.publication.reports.util.DateUtil._
@@ -55,52 +54,17 @@ object D51 {
     val msa = msaReport(fipsCode.toString)
 
     val incomeIntervals = calculateMedianIncomeIntervals(fipsCode)
-
-    val larsByIncome = larsByIncomeInterval(larsWithIncome, incomeIntervals)
-    val borrowerCharacteristicsByIncomeF = borrowerCharacteristicsByIncomeInterval(larsByIncome, dispositions)
+    val applicantIncomesF = applicantIncomesWithBorrowerCharacteristics(larsWithIncome, incomeIntervals, dispositions)
 
     val yearF = calculateYear(larSource)
     val totalF = calculateDispositions(lars, dispositions)
 
     for {
-      lars50BorrowerCharacteristics <- borrowerCharacteristicsByIncomeF(LessThan50PercentOfMSAMedian)
-      lars50To79BorrowerCharacteristics <- borrowerCharacteristicsByIncomeF(Between50And79PercentOfMSAMedian)
-      lars80To99BorrowerCharacteristics <- borrowerCharacteristicsByIncomeF(Between80And99PercentOfMSAMedian)
-      lars100To120BorrowerCharacteristics <- borrowerCharacteristicsByIncomeF(Between100And119PercentOfMSAMedian)
-      lars120BorrowerCharacteristics <- borrowerCharacteristicsByIncomeF(GreaterThan120PercentOfMSAMedian)
-
       institutionName <- institutionNameF
       year <- yearF
+      applicantIncomes <- applicantIncomesF
       total <- totalF
     } yield {
-      val income50 = ApplicantIncome(
-        LessThan50PercentOfMSAMedian,
-        lars50BorrowerCharacteristics
-      )
-      val income50To79 = ApplicantIncome(
-        Between50And79PercentOfMSAMedian,
-        lars50To79BorrowerCharacteristics
-      )
-      val income80To99 = ApplicantIncome(
-        Between80And99PercentOfMSAMedian,
-        lars80To99BorrowerCharacteristics
-      )
-      val income100To120 = ApplicantIncome(
-        Between100And119PercentOfMSAMedian,
-        lars100To120BorrowerCharacteristics
-      )
-      val income120 = ApplicantIncome(
-        GreaterThan120PercentOfMSAMedian,
-        lars120BorrowerCharacteristics
-      )
-
-      val applicantIncomes = List(
-        income50,
-        income50To79,
-        income80To99,
-        income100To120,
-        income120
-      )
 
       D51(
         respondentId,
